RR-173: Fixed code review misunderstanding
