+++ src/apps/events/client/Form/EventForm.jsx
@@ -10,12 +10,7 @@ import { Main } from '../../../../client/components'
 import TaskForm from '../../../../client/components/Task/Form'
 import { TASK_GET_EVENTS_FORM_AND_METADATA, TASK_SAVE_EVENT } from './state'
 import { EventFormFields } from './EventFormFields'
-import {
-  transformDateObjectToDateString,
-  transformArrayOfUniqueOptions,
-  transformOption,
-  transformYesNoToBool,
-} from '../../../../client/transformers'
+import { transformEventFormForAPIRequest } from './transformers'
 
 const DISPLAY_EDIT_EVENT = 'Edit event'
 const DISPLAY_ADD_EVENT = 'Add event'
@@ -39,52 +34,6 @@ const EventForm = ({ eventId }) => {
       text: eventId ? DISPLAY_EDIT_EVENT : DISPLAY_ADD_EVENT,
     },
   ]
-
-  const transformPayloadForEventForm = (values) => {
-    const {
-      has_related_trade_agreements,
-      related_trade_agreements,
-      start_date,
-      end_date,
-      event_type,
-      location_type,
-      address_country,
-      lead_team,
-      service,
-      organiser,
-      event_shared,
-      related_programmes,
-      teams,
-      uk_region,
-    } = values
-    const hasRelatedTradeAgreements = transformYesNoToBool(
-      has_related_trade_agreements
-    )
-    const hasEventShared = transformYesNoToBool(event_shared)
-    const transformedValuesOnlyPayload = {
-      has_related_trade_agreements: hasRelatedTradeAgreements,
-      related_trade_agreements: hasRelatedTradeAgreements
-        ? transformArrayOfUniqueOptions(related_trade_agreements)
-        : [],
-      event_type: transformOption(event_type),
-      start_date: transformDateObjectToDateString(start_date),
-      end_date: transformDateObjectToDateString(end_date),
-      location_type: transformOption(location_type),
-      address_country: transformOption(address_country),
-      lead_team: transformOption(lead_team),
-      service: transformOption(service),
-      organiser: transformOption(organiser),
-      event_shared: transformYesNoToBool(event_shared),
-      related_programmes: transformArrayOfUniqueOptions(related_programmes),
-      teams: hasEventShared
-        ? transformArrayOfUniqueOptions(
-            teams ? teams.concat([lead_team]) : [lead_team]
-          )
-        : transformArrayOfUniqueOptions([lead_team]),
-      uk_region: transformOption(uk_region),
-    }
-    return { ...values, ...transformedValuesOnlyPayload }
-  }
 
   const cancelLink = [
     {
@@ -120,7 +69,7 @@ const EventForm = ({ eventId }) => {
               flashMessage={flashMessage}
               submitButtonLabel={eventId ? DISPLAY_SAVE : DISPLAY_ADD_EVENT}
               actionLinks={cancelLink}
-              transformPayload={transformPayloadForEventForm}
+              transformPayload={transformEventFormForAPIRequest}
             >
               {({ values }) => <EventFormFields values={values} />}
             </TaskForm>
